+++ src/Framework/ClassResolver/ClassNameFinder/ClassNameFinder.php
@@ -5,35 +5,29 @@ declare(strict_types=1);
 namespace Gacela\Framework\ClassResolver\ClassNameFinder;
 
 use Gacela\Framework\ClassResolver\ClassInfo;
+use Gacela\Framework\ClassResolver\ClassNameCacheInterface;
 use Gacela\Framework\ClassResolver\ClassNameFinder\Rule\FinderRuleInterface;
 
 final class ClassNameFinder implements ClassNameFinderInterface
 {
-    /** @var array<string,string> */
-    private static array $cachedClassNames = [];
-
     private ClassValidatorInterface $classValidator;
 
     /** @var list<FinderRuleInterface> */
     private array $finderRules;
+
+    private ClassNameCacheInterface $classNameCache;
 
     /**
      * @param list<FinderRuleInterface> $finderRules
      */
     public function __construct(
         ClassValidatorInterface $classValidator,
-        array $finderRules
+        array $finderRules,
+        ClassNameCacheInterface $classNameCache
     ) {
         $this->classValidator = $classValidator;
         $this->finderRules = $finderRules;
-    }
-
-    /**
-     * @internal
-     */
-    public static function resetCachedClassNames(): void
-    {
-        self::$cachedClassNames = [];
+        $this->classNameCache = $classNameCache;
     }
 
     /**
@@ -43,15 +37,16 @@ final class ClassNameFinder implements ClassNameFinderInterface
     {
         $cacheKey = $classInfo->getCacheKey();
 
-        if (isset(self::$cachedClassNames[$cacheKey])) {
-            return self::$cachedClassNames[$cacheKey];
+        if ($this->classNameCache->has($cacheKey)) {
+            return $this->classNameCache->get($cacheKey);
         }
 
         foreach ($this->finderRules as $finderRule) {
             foreach ($resolvableTypes as $resolvableType) {
                 $className = $finderRule->buildClassCandidate($classInfo, $resolvableType);
                 if ($this->classValidator->isClassNameValid($className)) {
-                    self::$cachedClassNames[$cacheKey] = $className;
+                    $this->classNameCache->put($cacheKey, $className);
+
                     return $className;
                 }
             }